+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Now test cases can be asynchronous
 * toNotBe assertion
 * Optimisation for better performance
+* Hotkeys for watch mode
 
 #### Security Updates:
 * Updated `micromatch` from 4.0.7 to 4.0.8